typecheck_decorator: redefined IterableChecker to now be TupleChecker
IterableChecker could formerly be used on strings and would then recurse infinitely.
It could be used on sets where item order is undefined (and so a specification makes only a very little sense)
It could be used dictionaries where item order is undefined and only the names of the items would have been checked.
TupleChecker is now much more specific and no longer prone to weird uses.